+++ x/go/gorp/update_test.go
@@ -30,25 +30,25 @@ var _ = Describe("update", Ordered, func() {
 		kv = memkv.New()
 		db = gorp.Wrap(kv)
 		for i := range 10 {
-			entries = append(entries, entry{ID: i, Data: "data"})
+			entries = append(entries, entry{ID: int32(i), Data: "data"})
 		}
 	})
 	AfterAll(func() { Expect(kv.Close()).To(Succeed()) })
 	BeforeEach(func() {
 		tx = db.OpenTx()
-		Expect(gorp.NewCreate[int, entry]().Entries(&entries).Exec(ctx, tx)).To(Succeed())
+		Expect(gorp.NewCreate[int32, entry]().Entries(&entries).Exec(ctx, tx)).To(Succeed())
 	})
 	AfterEach(func() { Expect(tx.Close()).To(Succeed()) })
 
 	It("Should correctly update set of entries", func() {
-		Expect(gorp.NewUpdate[int, entry]().
+		Expect(gorp.NewUpdate[int32, entry]().
 			WhereKeys(entries[0].GorpKey()).
 			Change(func(_ gorp.Context, e entry) entry {
 				e.Data = "new data"
 				return e
 			}).Exec(ctx, tx)).To(Succeed())
 		var res entry
-		Expect(gorp.NewRetrieve[int, entry]().
+		Expect(gorp.NewRetrieve[int32, entry]().
 			WhereKeys(entries[0].GorpKey()).
 			Entry(&res).
 			Exec(ctx, tx)).To(Succeed())
@@ -56,13 +56,13 @@ var _ = Describe("update", Ordered, func() {
 	})
 
 	It("Should return an error if no change function was specified", func() {
-		Expect(gorp.NewUpdate[int, entry]().
+		Expect(gorp.NewUpdate[int32, entry]().
 			WhereKeys(entries[0].GorpKey()).
 			Exec(ctx, tx)).To(testutil.HaveOccurredAs(query.InvalidParameters))
 	})
 
 	It("Should return an error if the the key cannot be found", func() {
-		Expect(gorp.NewUpdate[int, entry]().
+		Expect(gorp.NewUpdate[int32, entry]().
 			WhereKeys(999).
 			Change(func(_ gorp.Context, e entry) entry {
 				e.Data = "new data"
@@ -72,7 +72,7 @@ var _ = Describe("update", Ordered, func() {
 
 	It("Should pass the correct transaction into the gorp.Context in the where function", func() {
 		count := 0
-		Expect(gorp.NewUpdate[int, entry]().
+		Expect(gorp.NewUpdate[int32, entry]().
 			WhereKeys(entries[0].GorpKey()).
 			Change(func(uCtx gorp.Context, e entry) entry {
 				e.Data = "new data"
@@ -86,25 +86,25 @@ var _ = Describe("update", Ordered, func() {
 
 	Describe("Where", func() {
 		It("Should correctly update a set of entries based on a where filter function", func() {
-			Expect(gorp.NewUpdate[int, entry]().
+			Expect(gorp.NewUpdate[int32, entry]().
 				Where(func(ctx gorp.Context, e *entry) (bool, error) { return e.ID < 5, nil }).
 				Change(func(_ gorp.Context, e entry) entry {
 					e.Data = "new data"
 					return e
 				}).Exec(ctx, tx)).To(Succeed())
 			var res []entry
-			Expect(gorp.NewRetrieve[int, entry]().
+			Expect(gorp.NewRetrieve[int32, entry]().
 				Where(func(ctx gorp.Context, e *entry) (bool, error) { return e.ID < 5, nil }).
 				Entries(&res).
 				Exec(ctx, tx)).To(Succeed())
 			for i := range res {
-				Expect(res[i]).To(Equal(entry{ID: i, Data: "new data"}))
+				Expect(res[i]).To(Equal(entry{ID: int32(i), Data: "new data"}))
 			}
 		})
 
 		It("Should pass the correct transaction to the gorp.Context in the where function", func() {
 			count := 0
-			Expect(gorp.NewUpdate[int, entry]().
+			Expect(gorp.NewUpdate[int32, entry]().
 				WhereKeys(entries[0].GorpKey()).
 				Where(func(wCtx gorp.Context, e *entry) (bool, error) {
 					count++
